refactor: Update clinic response to use FullAddress instead of Address
The `GetClinicsQueryHandler` was modified to use `FullAddress` property instead of `Address` when constructing the clinic response. This change aligns with the domain model where the address information is now stored in a more comprehensive `FullAddress` field. No breaking changes introduced as this is an internal implementation detail of the response mapping.